requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
-r requirements/production.txt

Django==2.0.1
django-crispy-forms==1.7.0
django-extensions==1.9.8
django-filter==1.1.0
django-model-utils==3.0.0
django-registration-redux==2.0
django-tables2==1.17.1
pytz==2017.3
six==1.11.0
typing==3.6.2


